Coursera Capstone - Upload
PGA Tour Analysis - Coursera Capstone

This is the final project in IBM's Introduction to Data Science. 

Using a kaggle dataset, this analysis dives into the metrics of golf. The dataset contains statistics from players on the PGA Tour between 2010 and 2018. In this report, we introduce the topic and the problem, explore the dataset and visualize correlations, and finally build a Linear Regression model as a predictor of success. The target audience of this report is large sports or Golf brands, such as Nike or Taylormade, as they search for their next PGA pro to sponsor. Studies have shown that athletes can have a tremendous effect on product sales, and brands are always looking for the top players to sponsor. This analysis will also have relevance to amateur golfers trying to improve their game, or even fans of the sport.